requirements.txt
Flask==3.0.2
PyJWT==2.8.0
requests==2.31.0